Fix for CR-1150866 , CR-1150870
	- Incorrectly reading Sysmon max temp , changed it to read sysmon current temp.
	- Moving average sum changed to float to aviod type casting issues.
	- Assigned 4bytes for vccint current reading , to send it through ASDM.

+++ vmr/src/vmc/clock_throttling.c
@@ -33,7 +33,7 @@ float Moving_Average(Moving_Average_t * av_obj, float new_element)
 		av_obj->is_filled = true;
 	}
 	/* return the average */
-	return av_obj->sum / (av_obj->is_filled ? av_obj->length:av_obj->pos);
+	return av_obj->sum / (float)(av_obj->is_filled ? av_obj->length:av_obj->pos);
 }
 
 Moving_Average_t *Allocate_Moving_Average(u8 len)
